module creation: #league_statable.rb is created and pulled league methods into the module
