+++ spec/stat_tracker_spec.rb
@@ -22,20 +22,17 @@ RSpec.describe StatTracker do
 
   describe "#initialize" do 
     it "exists" do 
-
       expect(stat_tracker).to be_a StatTracker
     end
   end
 
   describe "Game Statistics" do 
     it "#highest_total_score" do 
-    # require 'pry';binding.pry
       expect(stat_tracker.highest_total_score).to be_a(Integer)
       expect(stat_tracker.highest_total_score).to eq(11)
     end
 
     it "#lowest_total_score" do 
-    # require 'pry';binding.pry
       expect(stat_tracker.lowest_total_score).to be_a(Integer)
       expect(stat_tracker.lowest_total_score).to eq(0)
     end
